Split the worker and storage into their own files
Remove unneeded require statements
